preprocess_service/requirements.txt
uvicorn~=0.23.2
fastapi~=0.99.1
pillow~=9.4.0
numpy~=1.25.2
torchvision~=0.15.2